// ui/src/generated/ApplicationChanges.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { ApplicationContextUpdate } from "./ApplicationContextUpdate";
import type { ApplicationDocumentRef } from "./ApplicationDocumentRef";
import type { ApplicationDocumentUpdate } from "./ApplicationDocumentUpdate";

export type ApplicationChanges = { context: ApplicationContextUpdate | null, documents: Array<ApplicationDocumentUpdate>, removed_documents: Array<ApplicationDocumentRef>, };