// ui/src/generated/ApplicationSyncReceipt.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { ApplicationDocumentRef } from "./ApplicationDocumentRef";

export type ApplicationSyncReceipt = { sync_id: string, synced_at_ms: number, context_version: string, document_versions: Array<ApplicationDocumentRef>, };